requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tensorflow>=2.12
gensim>=4.3
PyYAML>=6.0
matplotlib>=3.7
joblib>=1.3
